Admin Panel Finished

+++ resources/views/admin/categories/index.blade.php
@@ -38,39 +38,42 @@
               </td>
             </tr>
           @empty
-
+            <h3>Данные отсутствуют</h3>
           @endforelse
         </tbody>
       </table>
-      <div class="modal fade" id="trashModal">
-        <div class="modal-dialog">
-          <div class="modal-content">
-            <div class="modal-header"><span class="text-danger">Внимание!</span></div>
-            <div class="modal-body">
-              При удалении этой категории все подкатегории и продукты будут удалены.
-            </div>
-            <div class="modal-footer">
-              <form method="POST" id="forceDelete" action="{{route('admin.categories.destroy', ["category" => $category->id + 1])}}">  
-                @csrf
-                @method('DELETE')
-                <input type="hidden" name="id" />
-                <button class="btn btn-danger"><i class="fa fa-trash"></i> Удалить все</button>
-              </form>
-              <button type="button" class="btn btn-default" data-dismiss="modal">Отмена</button>
+      @if (isset($category))
+        <div class="modal fade" id="trashModal">
+          <div class="modal-dialog">
+            <div class="modal-content">
+              <div class="modal-header"><span class="text-danger">Внимание!</span></div>
+              <div class="modal-body">
+                При удалении этой категории все подкатегории и продукты будут удалены.
+              </div>
+              <div class="modal-footer">
+                    <form method="POST" id="forceDelete" action="{{route('admin.categories.destroy', ["category" => $category->id + 1])}}">     
+                    <form method="POST" id="forceDelete" action="{{route('admin.categories.destroy', ["category" => $category->id + 1])}}">                         
+                  @csrf
+                  @method('DELETE')
+                  <input type="hidden" name="id" />
+                  <button class="btn btn-danger"><i class="fa fa-trash"></i> Удалить все</button>
+                </form>
+                <button type="button" class="btn btn-default" data-dismiss="modal">Отмена</button>
+              </div>
             </div>
           </div>
         </div>
-      </div>
+      @push('scripts')
+          <script>
+            $(document).on('click', '.del', function(){
+              const 
+                id = $(this).data('id'),
+                action = $("#forceDelete").attr('action').split('/');
+                action[action.length - 1] = id;
+              $("#forceDelete").attr('action', action.join('/'));
+            });
+          </script>    
+      @endpush
+    @endif
     </div>
-    @push('scripts')
-        <script>
-          $(document).on('click', '.del', function(){
-            const 
-              id = $(this).data('id'),
-              action = $("#forceDelete").attr('action').split('/');
-              action[action.length - 1] = id;
-            $("#forceDelete").attr('action', action.join('/'));
-          });
-        </script>    
-    @endpush
 @endsection
